lto, workaround: Disable LTO for BPF
Disable LTO for the BPF interpreter. This works around a gcc bug in the LTO
partitioner that partitions the jumptable from the BPF interpreter
into a different LTO unit. This in term causes assembler
errors because the jump table contains references to the
code labels in the original file.

gcc problem tracked in https://gcc.gnu.org/bugzilla/show_bug.cgi?id=50676

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: celtare21 <celtare21@gmail.com>