msm: mdss: optimize DSI command delay semantics
Some DSI commands require a delay to occur between them and the next
command sent to the panel. The existing code supports this, but is
not optimal since it sleeps *after* sending the command requiring the
delay, rather than immediately *before* the next command.

While this doesn't really matter if the command requiring the delay
is in the middle of a sequence of commands, it can make quite a big
difference in scenarios where the last command to be executed for a
while has a delay associated with it.

A good example is panel unblank/power-up sequences, which often end
with a command requiring a large delay (ex. 120 ms). Since the next
command to execute may not occur for a very long time, there should
be no need to sleep for a full 120 ms.

To improve this, and display turn-on latency as a result, note the
time at which the next command is allowed to be sent. Prior to
sending the next command, compare this time with the current time.
If the required amount of time has already elapsed, skip sleeping
entirely; otherwise, sleep for the difference.

Change-Id: Ibfe723267dd0abd8f5ebc8cb7bf0c17c6fa640c0
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>
Signed-off-by: celtare21 <celtare21@gmail.com>