FROMLIST: uaccess: add untagged_addr definition for other arches
(from https://lore.kernel.org/patchwork/patch/994345/)

To allow arm64 syscalls accept tagged pointers from userspace, we must
untag them when they are passed to the kernel. Since untagging is done in
generic parts of the kernel, the untagged_addr macro needs to be defined
for all architectures.

Define it as a noop for other architectures besides arm64.

Bug: 112461694
Change-Id: I7b9bd43d4f27492f49b4374d3aec96577f44b7fb
Signed-off-by: Andrey Konovalov <andreyknvl@google.com>
Signed-off-by: celtare21 <celtare21@gmail.com>