input: hbtp: Hack around pressure reporting bug
If you touch the screen with 2 or more fingers at the same time,
anything that is not finger[1] will be bugged for the first touch
and report a constant value of 255. This wouldn't be a problem with
the original very low pressure values, but we're inflating the value
by 8 times to get proper readins. Because of this, hack around the bug
so the inflated value doesn't report an insane ~2000 pressure value.
We do this by:
a) Always having a sane fallback value being the last value read
   before the broken value.
b) If this fails, have a sanity check and just reject the broken
   value anyways.

Signed-off-by: celtare21 <celtare21@gmail.com>